+++ src/ConfigEditor.tsx
@@ -1,7 +1,8 @@
 import React, { ChangeEvent, PureComponent } from 'react';
-import { DataSourceHttpSettings } from '@grafana/ui';
-import { DataSourcePluginOptionsEditorProps } from '@grafana/data';
+import { DataSourceHttpSettings, LegacyForms} from '@grafana/ui';
+import { DataSourcePluginOptionsEditorProps, onUpdateDatasourceJsonDataOptionChecked } from '@grafana/data';
 import { SGApiDataSourceOptions } from './types';
+const { Switch } = LegacyForms;
 
 interface Props extends DataSourcePluginOptionsEditorProps<SGApiDataSourceOptions> {}
 
@@ -12,7 +13,7 @@ export class ConfigEditor extends PureComponent<Props, State> {
     const { onOptionsChange, options } = this.props;
     const jsonData = {
       ...options.jsonData,
-      path: event.target.value,
+      showOfflineDatasources: event.target.checked
     };
     onOptionsChange({ ...options, jsonData });
   };
@@ -29,8 +30,27 @@ export class ConfigEditor extends PureComponent<Props, State> {
             showAccessOptions={true}
             //sigV4AuthEnabled={true}
             onChange={onOptionsChange}
-          />  
+          />
         </div>
+        <div className="gf-form">
+          <div className="gf-form-group">
+            <h3 className="page-heading">Misc</h3>
+            <div className="gf-form-group">
+
+              <div className="gf-form">
+                <Switch
+                  checked={options.jsonData.showOfflineDatasources ?? false}
+                  label="Show offline Datasources" 
+                  labelClass="width-14"
+                  onChange={onUpdateDatasourceJsonDataOptionChecked(this.props, 'showOfflineDatasources')}
+                  tooltip="Check this option to include temporary offline data sources in the query editor"
+                />
+              </div>
+
+            </div>
+          </div>
+        </div>
+
       </div>
     );
   }